# Optimize `invalidDocOffsets(_:)`