Change guard condition to fails earlier

This commit reduces the duration of linting KeychainAccess from:
```
swiftlint lint  5.37s user 0.14s system 93% cpu 5.889 total
```
to:
```
swiftlint lint  3.74s user 0.13s system 91% cpu 4.238 total
```

the duration of linting Carthage is reduced from:
```
swiftlint lint  33.48s user 1.86s system 82% cpu 42.645 total
```
to:
```
swiftlint lint  29.61s user 1.81s system 79% cpu 39.377 total
```