@jpsim 2015-12-23
This is great! I didn't even know you could have multiple `where` conditions! I learned something very useful today! :blush: