@norio-nomura 2015-12-24
I think there is room for more performance optimization in SwiftLint and SourceKitten, but that will require breaking changes on SourceKitten and I don't know how that affects to another projects.
And I had read [the post of Realm Blog](https://realm.io/news/swift-opensource/#open-source-swift--realm).

I think I should wait to doing them until the new action appears that hinted in the post. :smiley: