@jpsim 2015-12-24
Yeah, there's definitely more room for performance improvements. I don't know if anything we said in the blog post will lead to perf improvements though! I'd expect that using SourceKit in process rather than through XPC calls would speed things up a bit, but probably not all that much.