@norio-nomura 2015-12-24
Yes, I expect using SourceKit in process.

By the way, now I'm surprised that swift-2.2 produce faster binaries.:zap:
I builded SwiftLint 26dba4e (with some patch to SourceKitten) using swift-2.2-SNAPSHOT-2015-12-22-a.
the duration of linting KeychainAccess is:
```
swiftlint lint  2.71s user 0.12s system 83% cpu 3.395 total
```
and linting Carthage is:
```
swiftlint lint  21.35s user 1.67s system 77% cpu 29.687 total
```
I didn't think about performance improvements of Swift-2.2.