@jpsim 2015-12-24
Whoa... those performance gains from Swift 2.2 are crazy! I can't wait! :smile: