@norio-nomura 2015-12-25
FYI, [How to use swift-2.2-SNAPSHOT in xcodebuild](https://gist.github.com/norio-nomura/ea63b61ab942fff2e2b3)

On building SwiftLint, it needs patch for SourceKitten as following:
```sh
sed -i "~" "s/@asmname/@_silgen_name/" Carthage/Checkouts/SourceKitten/Source/SourceKittenFramework/sourcekitd.swift
```
then it can be built by following:
```sh
XCODE_XCCONFIG_FILE=/path/to/swift-latest.xcconfig make install
```

It needs `swift-2.2-SNAPSHOT-2015-12-22-a` or later because that contains fix of [a bug SR-329](https://bugs.swift.org/browse/SR-329).